New map base, new controller.
Discarded the previous map as I didn't like its structure and began creating a new one. The map will stay pretty basic for now as I want to focus on improving the character controller and also create a climbing system. Hopefully I will manage to finish the new 3rd person controller till the end of autumn break. As of now, the animator's structure is more or less finished.